% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcmdr.R
\name{just_Rcmdr}
\alias{just_Rcmdr}
\title{Load R Commander alone. This way, run Rcmdr()
and the environment will be ready.}
\usage{
just_Rcmdr()
}
\description{
Load R Commander alone. This way, run Rcmdr()
and the environment will be ready.
}
\examples{
just_Rcmdr()
}

