% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{pval}
\alias{pval}
\title{Run the P-val Walkthrough script.}
\usage{
pval()
}
\value{
Acceptable user input.
}
\description{
Run the P-val Walkthrough script.
}
\examples{
\dontrun{pval()}
}
\author{
Josh Errickson
}

