% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_shinies.R
\name{s_introbinom}
\alias{s_introbinom}
\title{Run the Intro Binomial shiny app.
 @export}
\usage{
s_introbinom()
}
\description{
Run the Intro Binomial shiny app.
 @export
}
\examples{
\dontrun{s_introbinom()}
}

