% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/call_shinies.R
\name{s_propci}
\alias{s_propci}
\title{Run the proportion confidence interval shiny app.}
\usage{
s_propci()
}
\description{
Run the proportion confidence interval shiny app.
}
\examples{
s_propci()
}

