% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/call_shinies.R
\name{s_twonorm}
\alias{s_twonorm}
\title{Run the Two Normal Distributions shiny app.}
\usage{
s_twonorm()
}
\description{
Run the Two Normal Distributions shiny app.
}
\examples{
\dontrun{s_twonorm()}
}

