% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriessim.R
\name{timeseressim}
\alias{timeseressim}
\title{Run the Time Series Simulation script.}
\usage{
timeseressim()
}
\description{
Run the Time Series Simulation script.
}
\examples{
\dontrun{timeseriessim()}
}
\author{
Omar Chavez
}

