% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/call_shinies.R
\name{s_normbinom}
\alias{s_normbinom}
\title{Run the Normal Approximation to the Binomial shiny app.}
\usage{
s_normbinom()
}
\description{
Run the Normal Approximation to the Binomial shiny app.
}
\examples{
\dontrun{s_normbinom()}
}

