% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/timeseriessim.R
\name{timeseressim}
\alias{timeseressim}
\title{Run the Time Series Simulation script.}
\usage{
timeseressim()
}
\description{
Run the Time Series Simulation script.
}
\examples{
\dontrun{timeseriessim()}
}

